// star-kitten-lib/src/zORM/types.ts
export type PropertyGetter<T> = (obj: T) => any;
export type MultiPropertyGetter<T> = (obj: T) => any[];

export type SQliteTypes = 'TEXT' | 'INTEGER' | 'DECIMAL' | 'BLOB' | 'NULL' | 'BIGINT';

export enum DataType {
  BIGINT = 'BIGINT',
  STRING = 'TEXT',
  INTEGER = 'INTEGER',
  DECIMAL = 'DECIMAL',
  REAL = 'FLOAT',
  BLOB = 'BLOB',
  NULL = 'NULL',
}

export type ColumnMetaDataType = 'column' | 'one-to-many' | 'one-to-one';
export type VirtualColumnMetaDataType = 'join-one' | 'join-many' | 'many-to-one';
export type IndexType = 'index' | 'unique-index' | 'composite-index';
export type MetaDataType = 'table' | ColumnMetaDataType | IndexType | VirtualColumnMetaDataType;

export interface HasMetaData {
  key: string;
  classname: string;
  type: MetaDataType;
  options?:
  | ColumnDecoratorOptions
  | RelationMetaDataOptions
  | TableDecoratorOptions
  | (ManyToOneDecoratorOptions & RelationMetaDataOptions)
  | IndexMetaDataOptions
  | CompositeIndexMetaDataOptions
  | MultiRelationMetaDataOptions;
}
// ...
export interface TableDecoratorOptions {
  name?: string;
  DATABASE?: string;
  indices?: PropertyGetter<any>[];
  uniqueIndices?: PropertyGetter<any>[];
  compositeIndices?: MultiPropertyGetter<any>[];
}

export interface ColumnDecoratorOptions {
  name?: string; // name of the column in the database
  type?: DataType; // type of the column in the database
  primary?: boolean; // if the column is a primary key
  unique?: boolean; // if the column is unique
  nullable?: boolean; // if the column is nullable
}

export interface RelationMetaDataOptions {
  referencedTable: string;
  referencedKey: string;
}

export interface MultiRelationMetaDataOptions {
  referencedTable: string;
  referencedKeys: string[][];
}

export interface ManyToOneDecoratorOptions {
  cascade?: boolean;
}

export interface IndexMetaDataOptions {
  keys: string[];
}

export interface CompositeIndexMetaDataOptions {
  keys: string[][];
}